Rename recompute_bin_data_c -> compute_pix_sums_c
This is more descriptive of what the C++ routine is actually doing. It's
not calculating a average anymore, it's only calculating a sum
