+++ _LowLevelCode/cpp/compute_pix_sums/compute_pix_sums_helpers.cpp
@@ -1,4 +1,4 @@
-#include "recompute_bin_data/recompute_bin_data_helpers.h"
+#include "compute_pix_sums/compute_pix_sums_helpers.h"
 
 #include <memory>
 #include <sstream>
@@ -7,7 +7,7 @@ void validate_inputs(const int &nlhs, mxArray *plhs[], const int &nrhs,
                      const mxArray *prhs[]) {
   if (nrhs != N_INPUT_Arguments && nrhs != N_INPUT_Arguments - 1) {
     std::stringstream buf;
-    buf << "ERROR::recompute_bin_data_c needs " << (short)N_INPUT_Arguments
+    buf << "ERROR::compute_pix_sums_c needs " << (short)N_INPUT_Arguments
         << " but got " << (short)nrhs << " input arguments and " << (short)nlhs
         << " output argument(s)\n";
     mexErrMsgTxt(buf.str().c_str());
@@ -15,7 +15,7 @@ void validate_inputs(const int &nlhs, mxArray *plhs[], const int &nrhs,
 
   if (nlhs != N_OUTPUT_Arguments) {
     std::stringstream buf;
-    buf << "ERROR::recompute_bin_data_c needs " << (short)N_OUTPUT_Arguments
+    buf << "ERROR::compute_pix_sums_c needs " << (short)N_OUTPUT_Arguments
         << " outputs but requested to return" << (short)nlhs << " arguments\n";
     mexErrMsgTxt(buf.str().c_str());
   }
@@ -23,7 +23,7 @@ void validate_inputs(const int &nlhs, mxArray *plhs[], const int &nrhs,
   for (int i = 0; i < nrhs - 1; i++) {
     if (prhs[i] == NULL) {
       std::stringstream buf;
-      buf << "ERROR::recompute_bin_data_c => argument N" << i << " undefined\n";
+      buf << "ERROR::compute_pix_sums_c => argument N" << i << " undefined\n";
       mexErrMsgTxt(buf.str().c_str());
     }
   }
@@ -32,7 +32,7 @@ void validate_inputs(const int &nlhs, mxArray *plhs[], const int &nrhs,
 const double *const get_npix_array(const mxArray *prhs[]) {
   const double *const p_npix_data = (double *)mxGetPr(prhs[Npix_data]);
   if (!p_npix_data) {
-    mexErrMsgTxt("ERROR::recompute_bin_data_c-> undefined or empty npix array");
+    mexErrMsgTxt("ERROR::compute_pix_sums_c-> undefined or empty npix array");
   }
 
   return p_npix_data;
@@ -42,14 +42,14 @@ const double *const get_pixel_array(const mxArray *prhs[]) {
   // Validate number of pixel data columns
   std::size_t nPixDataCols = mxGetM(prhs[Pixel_data]);
   if (nPixDataCols != pix_fields::PIX_WIDTH) {
-    mexErrMsgTxt("ERROR::recompute_bin_data_c-> the pixel data should be a "
+    mexErrMsgTxt("ERROR::compute_pix_sums_c-> the pixel data should be a "
                  "9*num_of_pixels array");
   }
 
   const double *const p_pixel_data = (double *)mxGetPr(prhs[Pixel_data]);
   if (!p_pixel_data) {
     mexErrMsgTxt(
-        "ERROR::recompute_bin_data_c-> undefined or empty pixels array");
+        "ERROR::compute_pix_sums_c-> undefined or empty pixels array");
   }
 
   return p_pixel_data;
@@ -69,7 +69,7 @@ double *get_output_signal_ptr(mwSize &num_dims, const mwSize *dims,
                               mxArray *plhs[]) {
   plhs[Signal] = mxCreateNumericArray(num_dims, dims, mxDOUBLE_CLASS, mxREAL);
   if (!plhs[Signal]) {
-    mexErrMsgTxt("ERROR::recompute_bin_data_c-> can not allocate memory for "
+    mexErrMsgTxt("ERROR::compute_pix_sums_c-> can not allocate memory for "
                  "output signal array");
   }
   return (double *)mxGetPr(plhs[Signal]);
@@ -79,7 +79,7 @@ double *get_output_variance_ptr(mwSize &num_dims, const mwSize *dims,
                                 mxArray *plhs[]) {
   plhs[Variance] = mxCreateNumericArray(num_dims, dims, mxDOUBLE_CLASS, mxREAL);
   if (!plhs[Variance]) {
-    mexErrMsgTxt("ERROR::recompute_bin_data_c-> can not allocate memory for "
+    mexErrMsgTxt("ERROR::compute_pix_sums_c-> can not allocate memory for "
                  "output variance array");
   }
   return (double *)mxGetPr(plhs[Variance]);
