+++ horace_core/admin/check_horace_mex.m
@@ -7,7 +7,7 @@ function [rez, n_errors, can_use_mex_4_combine] = check_horace_mex()
 %   >>[rez, n_errors, can_use_mex_4_combine] = check_horace_mex();
 %
 % If some mex files are cannot be launched,the function returns the number of
-% files not launched as n_errors, these mex files' versions strings will be 
+% files not launched as n_errors, these mex files' versions strings will be
 % empty.
 %
 % rez is cellarray, which contains the reply from mex files queried about their
@@ -23,14 +23,14 @@ functions_name_list={
     'bin_pixels_c      : ', ...
     'calc_projections  : ', ...
     'sort_pixels_by_bin: ', ...
-    'recompute_bin_data: ', ...                                                   
+    'recompute_bin_data: ', ...
     'combine_sqw       : ', ...
     'mtimesx_mex       : ', ...
     'hdf_mex_reader    : '
    };
 
 % provide special treatment for combine_sqw function
-combine_num = numel(functions_name_list)-2; 
+combine_num = numel(functions_name_list)-2;
 
 % list of the mex file handles used by Horace and verified by this script.
 functions_handle_list={
@@ -38,7 +38,7 @@ functions_handle_list={
     @bin_pixels_c, ...
     @calc_projections_c, ...
     @sort_pixels_by_bins, ...
-    @recompute_bin_data_c, ...
+    @compute_pix_sums_c, ...
     @combine_sqw, ...
     @mtimesx_mex, ...
     @hdf_mex_reader
